+++ packages/io-lionweb-mps-specific/src/deserializer.ts
@@ -17,13 +17,13 @@
 
 import { LionWebJsonChunk } from "@lionweb/json"
 import {
-    defaultSimplisticHandler,
+    consoleProblemReporter,
     deserializerWith,
     Language,
     lioncore,
     lioncoreBuiltins,
     nodesExtractorUsing,
-    SimplisticHandler
+    ProblemReporter
 } from "@lionweb/core"
 import { lioncoreExtractionFacade } from "@lionweb/core/dist/m3/facade.js"
 import { ioLionWebMpsSpecificLanguage } from "./definition.js"
@@ -32,13 +32,13 @@ import { combinedWriter } from "./facade.js"
 /**
  * @return the deserialization of the given {@link LionWebJsonChunk serialization chunk} as an array of {@link Language languages}.
  * Any LionCore/M3 node can be annotated using annotations from the `io.lionweb.mps.specific` language.
- * Problems are reported through the given {@link SimplisticHandler} which defaults to {@link defaultSimplisticHandler}.
+ * Problems are reported through the given {@link ProblemReporter} which defaults to {@link consoleProblemReporter}.
  */
-export const deserializeLanguagesWithIoLionWebMpsSpecific = (serializationChunk: LionWebJsonChunk, problemHandler: SimplisticHandler = defaultSimplisticHandler) =>
+export const deserializeLanguagesWithIoLionWebMpsSpecific = (serializationChunk: LionWebJsonChunk, problemHandler: ProblemReporter = consoleProblemReporter) =>
     deserializerWith({
         writer: combinedWriter,
         languages: [lioncore, ioLionWebMpsSpecificLanguage],
-        problemHandler
+        problemReporter: problemHandler
     })(
         serializationChunk,
         [lioncore, lioncoreBuiltins].flatMap(nodesExtractorUsing(lioncoreExtractionFacade))